refactor(view): split the header's three jobs, trim the prose
