+++ crates/artifact/src/board/card.rs
@@ -8,31 +8,21 @@ pub struct Card {
     /// by the read that finds it missing — see [`super::Board::mint_ids`].
     #[serde(default)]
     pub id: String,
-    /// This card's number within its board — the `12` in `ROAD-12`, drawn from
-    /// [`super::Board::next_handle`] when the card is made.
-    ///
-    /// Apart from [`Card::id`] on purpose. The id is what the app and the tools
-    /// address a card by and it never has to be read aloud; the handle is what
-    /// a person says to an agent, so it is short, and being short it cannot
-    /// also be unique for all time. Defaulted for a card written before handles
-    /// existed — see [`super::Board::mint_ids`].
+    /// The `12` in `ROAD-12`: short enough to say to an agent, and so not
+    /// unique for all time the way [`Card::id`] is. Minted on read when
+    /// missing — see [`super::Board::mint_ids`].
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub handle: Option<u64>,
     pub text: String,
-    /// The session this card was dispatched to, by the id that session is
-    /// filed under.
-    ///
-    /// Persisted, unlike the run-local number it used to be: sessions are read
-    /// back from disk, so a link that died at quit left the board offering to
-    /// dispatch a card that already had an agent on it — and dispatching again
-    /// started a second one at the same task.
+    /// The session this card was dispatched to, by the id it is filed under.
+    /// Persisted, so ▶/💬 still tells the truth after a quit.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub session: Option<String>,
 }
 
 impl Card {
-    /// Under an id its board has found free — see [`super::Board::add_card`],
-    /// which is the only thing that knows what is taken.
+    /// Under an id and a handle its board has found free — see
+    /// [`super::Board::add_card`].
     pub fn new(id: String, handle: u64, text: String) -> Self {
         Self {
             id,
